finish editing

+++ quizz/models/answer.py
@@ -7,7 +7,7 @@ class Answer(models.Model):
     question = models.ForeignKey(Question, on_delete=models.CASCADE, related_name="answers", null=True, blank=True)
     answer = models.CharField(max_length=255, null=True, blank=True)
     is_correct = models.BooleanField(default=False)
-    points = models.IntegerField(blank=True, null=True)
+    points = models.FloatField(blank=True, null=True)
 
     def __str__(self):
         return self.answer
